Reorganise functions
