+++ _plugins/snippet.rb
@@ -135,20 +135,6 @@ module Jekyll
       end
     end
   end
-
-  module RegexReplace
-    def regex_replace(str, regex_search, value_replace)
-      regex = /#{regex_search}/m
-      return str.gsub(regex, value_replace)
-    end
-
-    def regex_replace_once(str, regex_search, value_replace)
-      regex = /#{regex_search}/m
-      return str.sub(regex, value_replace)
-    end
-  end
-
 end
 
 Liquid::Template.register_tag("snippet", Jekyll::Tags::SnippetIncludeTag)
-Liquid::Template.register_filter(Jekyll::RegexReplace)
